Move input to the new input system

+++ Assets/Scripts/Services/InputService/IInputCallback.cs
@@ -4,7 +4,7 @@ namespace Services.InputService
 {
     public interface IInputCallback
     {
-        event Action<float> OnHorizontalInput;
+        event Action<int> OnHorizontalInput;
         event Action OnJumpInput;
     }
 }
